+++ dandy/llm/config/ollama.py
@@ -14,8 +14,8 @@ class OllamaLlmConfig(BaseLlmConfig):
 
     def generate_request_body(
             self,
-            context_length: Union[int, None] = None,
-            max_completion_tokens: Union[int, None] = None,
+            max_input_tokens: Union[int, None] = None,
+            max_output_tokens: Union[int, None] = None,
             seed: Union[int, None] = None,
             temperature: Union[float, None] = None,
     ) -> BaseRequestBody:
@@ -23,8 +23,8 @@ class OllamaLlmConfig(BaseLlmConfig):
         return OllamaRequestBody(
             model=self.model,
             options=OllamaRequestOptions(
-                num_ctx=self.context_length if context_length is None else context_length,
-                num_predict=self.max_completion_tokens if max_completion_tokens is None else max_completion_tokens,
+                num_ctx=self.max_input_tokens if max_input_tokens is None else max_input_tokens,
+                num_predict=self.max_output_tokens if max_output_tokens is None else max_output_tokens,
                 seed=self.seed if seed is None else seed,
                 temperature=self.temperature if temperature is None else temperature
             )