+++ example/pirate/intelligence/configs.py
@@ -2,7 +2,6 @@ import os
 
 from dandy.llm.config import OllamaLlmConfig, OpenaiLlmConfig
 
-
 OLLAMA_LLAMA_3_2_3B = OllamaLlmConfig(
     host=os.getenv("OLLAMA_HOST"),
     port=int(os.getenv("OLLAMA_PORT", 11434)),
@@ -10,8 +9,8 @@ OLLAMA_LLAMA_3_2_3B = OllamaLlmConfig(
     model='llama3.2:3b-instruct-q4_K_M',
     temperature=1.0,
     prompt_retry_count=3,
-    context_length=8000,
-    max_completion_tokens=8000
+    max_input_tokens=8000,
+    max_output_tokens=8000
 )
 
 OLLAMA_LLAMA_3_1_8B = OllamaLlmConfig(
@@ -21,8 +20,8 @@ OLLAMA_LLAMA_3_1_8B = OllamaLlmConfig(
     model='llama3.1:8b-instruct-q4_K_M',
     temperature=1.0,
     prompt_retry_count=3,
-    context_length=8000,
-    max_completion_tokens = 8000
+    max_input_tokens=8000,
+    max_output_tokens=8000
 )
 
 OPENAI_GPT_3_5_TURBO = OpenaiLlmConfig(